backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
    "google-cloud-storage>=2.18.0",
    "pypdf>=5.0.0",
]
